# backend/api/app/static/swagger.json
{
	"openapi": "3.0.0",
	"servers": [
		{
			"description": "HydroTwin API",
			"url": "https://172.16.3.10:80/api"
		}
	],
	"info": {
		"description": "This swagger provides information of HydroTwin API",
		"version": "1.0.0",
		"title": "HydroTwin API",
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
...
	"paths": {
		"/api/saih/get-variable-values": {
			"post": {
				"tags": ["SAIH"],
				"summary": "Get SAIH variable values",
				"operationId": "get-values",
				"description": " By passing the variable code and a period of time you can retrieve its values.",
				"requestBody": {
					"$ref": "#/components/requestBodies/variableValueRequest"
				},
				"responses": {
					"200": {
						"description": "Return variables value in specified period of time.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/variableValueResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/demand-unit/get-by-position": {
			"post": {
				"tags": ["demand-unit"],
				"summary": "Load demand units neart to geojson geometry",
				"operationId": "demand-unit-get-by-position",
				"description": "By passing in the appropriate options, you can load all demand units near a geojson formated geometry.\n",
				"requestBody": {
					"$ref": "#/components/requestBodies/getByPosition"
				},
				"responses": {
					"200": {
						"description": "list of demand units near the geometry"
					},
					"400": {
						"description": "bad input parameter"
					},
					"500": {
						"description": "error during processing"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"variableValueRequest": {
				"required": ["start_date", "end_date", "variable"],
				"type": "object",
				"properties": {
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"variable": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"aggregation": {
						"type": "string"
					},
					"window": {
						"type": "string"
					}
				}
			},
			"variableValueResponse": {
				"type": "array",
				"items": {
					"type": "object",
					"properties": {
						"variableCode": {
							"type": "string"
						},
						"_time": {
							"type": "number"
						},
						"_value": {
							"type": "number"
						}
					}
				}
			},
...
			"getByPosition": {
				"type": "object",
				"required": ["geometry"],
				"properties": {
					"geometry": {
						"description": "GeoJson formated geometry",
						"type": "object",
						"$ref": "#/components/schemas/geometry"
					},
					"distance": {
						"description": "Maximun distance in meters, between geometry and the element. Default is 100 meters",
						"type": "integer",
						"example": 2000
					}
				}
			},
			"geometry": {
				"type": "object",
				"required": ["coordinates", "type"],
				"properties": {
					"type": {
						"description": "GeoJson geometry type",
						"type": "string",
						"example": "Point"
					},
					"coordinates": {
						"description": "GeoJson geometry coordinates bades on type",
						"type": "array",
						"items": {
							"type": "number"
						},
						"example": [-1.234375, 38.734375]
					}
				}
			}
		},
		"requestBodies": {
			"variableValueRequest": {
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/variableValueRequest"
						}
					}
				}
			},
...
			"getByPosition": {
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/getByPosition"
						}
					}
				}
			}
		}
	}
}